+++ CHANGELOG.md
@@ -21,7 +21,8 @@ Types of Changes:
 - The use of `multipart/form-data` on the POST `/api/hackers/` endpoint.
 
 ### Added
-
+-   POST `/api/sponsors/` endpoint for adding a hackathon sponsor to the DB.
+-   GET `/api/sponsors/get_all_sponsors/` endpoint for retrieving information on all of the hackathon sponsors.
 -   POST `/api/hackers/resume/` endpoint for uploading a hacker's resume before the hacker is actually created in the DB. Unattached resumes are deleted after 24 hours.
 -   Added the ability to upload the hacker document as a `application/json` while providing the optional `resume_id` value to attach a resume to a hacker document which is returned from the `/api/hackers/resume/` endpoint.